% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markdown_hfm.R
\name{create_markdown_hfm}
\alias{create_markdown_hfm}
\title{Generate Rmarkdown file}
\usage{
create_markdown_hfm(input_path)
}
\arguments{
\item{input_path}{A filepath of an excel table to import.}
}
\value{
Rmarkdown file.
}
\description{
Generate Rmarkdown file
}
